requirements.txt
yt-dlp
tk
ttkbootstrap
requests
importlib.metadata
pillow